requirements.txt
fastapi==0.111.0
matplotlib==3.9.0
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
scikit_learn==1.5.0
tensorflow==2.16.1
ultralytics==8.2.28
